requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
langchain_community
PyPDF2
chromadb
faiss-cpu
langchain_google_genai